+++ ItHappened.UnitTests/ServicesTests/EventServiceTests.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using ItHappened.Application.Errors;
 using ItHappened.Application.Services.EventService;
 using ItHappened.Domain;
-using ItHappened.Infrastructure.Repositories;
+using ItHappened.Infrastructure.InMemoryRepositories;
 using ItHappened.UnitTests.StatisticsCalculatorsTests;
 using LanguageExt;
 using NUnit.Framework;
