+++ ItHappened.UnitTests/StatisticsCalculatorsTests/MostFrequentEventCalculatorTest.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using ItHappened.Domain;
 using ItHappened.Domain.Statistics;
 using ItHappened.Infrastructure;
-using ItHappened.Infrastructure.Repositories;
+using ItHappened.Infrastructure.InMemoryRepositories;
 using LanguageExt.UnsafeValueAccess;
 using NUnit.Framework;
 using static ItHappened.UnitTests.StatisticsCalculatorsTests.StatisticsCalculatorsTestingConstants;
